% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_retention.R
\name{calculate_retention}
\alias{calculate_retention}
\title{Calculate Retention Rates}
\usage{
calculate_retention(data, intervals = seq(30, 450, 30))
}
\arguments{
\item{data}{The preprocessed comment data.}

\item{intervals}{A vector of time intervals in days.}
}
\value{
A data frame with retention rates.
}
\description{
This function calculates the retention rates of commenters over specified time intervals.
}
