% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group_data}
\alias{get_group_data}
\title{Extract Group Data Based on Initial Interaction}
\usage{
get_group_data(data, group_videos)
}
\arguments{
\item{data}{The preprocessed data frame containing comment data.}

\item{group_videos}{A vector of `Doc_name` values representing the group of interest (treated or control).}
}
\value{
A data frame containing all comments from commenters whose first comment was on a video in `group_videos`.
}
\description{
This function identifies commenters whose first comment was on a video in a specified group (e.g., treated or control),
and then retrieves all comments made by these commenters.
}
\examples{
\dontrun{
treated_data <- get_group_data(preprocessed_data, treated_group)
}
}
