% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_comment_data.R
\name{load_comment_data}
\alias{load_comment_data}
\title{Load YouTube Comment Data}
\usage{
load_comment_data(data_paths)
}
\arguments{
\item{data_paths}{A vector of file paths to the RData files.}
}
\value{
A combined data frame of all comments.
}
\description{
This function loads and combines multiple RData files containing YouTube comment data.
}
