% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_retention.R
\name{visualize_retention}
\alias{visualize_retention}
\title{Visualize Retention Rates}
\usage{
visualize_retention(treated_data, control_data)
}
\arguments{
\item{treated_data}{The data frame for the treated group.}

\item{control_data}{The data frame for the control group.}
}
\value{
A ggplot object.
}
\description{
This function generates a plot of retention rates with confidence intervals.
}
