% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_data_time.R
\name{get_group_data_time}
\alias{get_group_data_time}
\title{Extract Group Data Based on Initial Interaction within a Specified Time Frame}
\usage{
get_group_data_time(data, group_videos, days_after_release = 7)
}
\arguments{
\item{data}{The preprocessed data frame containing comment data.}

\item{group_videos}{A vector of `Doc_name` values representing the group of interest (treated or control).}

\item{days_after_release}{Number of days after video release within which the first comment must have been made (default is 7).}
}
\value{
A data frame containing all comments from commenters whose first comment was on a video in `group_videos` within the specified time frame.
}
\description{
This function identifies commenters whose first comment was on a video in a specified group (e.g., treated or control),
and who made their first comment within a specified number of days after the video release.
It then retrieves all comments made by these commenters.
}
\examples{
\dontrun{
treated_data <- get_group_data(preprocessed_data, treated_group, days_after_release = 7)
}
}
