% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_comments.R
\name{preprocess_comments}
\alias{preprocess_comments}
\title{Preprocess Comment Data}
\usage{
preprocess_comments(data)
}
\arguments{
\item{data}{The comment data frame.}
}
\value{
A preprocessed data frame.
}
\description{
This function preprocesses the comment data by adding necessary columns and filtering.
}
